backend/config.py
'''Contains all the configurations for the api'''
import os
from dotenv import load_dotenv


basedir = os.path.abspath(os.path.dirname(__file__))
load_dotenv(os.path.join(basedir, ".env"))


class Config(object):
    SQLALCHEMY_DATABASE_URI = 'sqlite:///' + os.path.join(basedir, 'cashx.db')
    SQLALCHEMY_TRACK_MODIFICATIONS = False
    SECRET_KEY = 'super top secret key'
    JWT_ACCESS_LIFESPAN = {'hours': 24}
    JWT_REFRESH_LIFESPAN = {'days': 30}
