backend/requirements.txt
alembic==1.7.4
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==7.1.2
Deprecated==1.2.13
finnhub-python==2.4.5
Flask==1.1.4
flask-buzz==0.1.15
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
flask-praetorian==1.2.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.5
MarkupSafe==2.0.1
passlib==1.7.4
pendulum==2.1.2
py-buzz==1.0.3
PyJWT==2.2.0
python-dateutil==2.8.2
python-dotenv==0.19.1
pytzdata==2020.1
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.25
urllib3==1.26.7
Werkzeug==1.0.1
wrapt==1.13.2
